refactor: code split
