+++ frontend/src/types/sqlEditor.ts
@@ -1,24 +1,12 @@
 import type * as monaco from "monaco-editor";
-import { InstanceId, DatabaseId, TableId, ViewId, ActivityId } from "../types";
+import { InstanceId, DatabaseId, ActivityId } from "../types";
 import { Principal } from "./principal";
 
 export type EditorModel = monaco.editor.ITextModel;
 export type EditorPosition = monaco.Position;
 export type CompletionItems = monaco.languages.CompletionItem[];
 
-export type ConnectionAtomType = "instance" | "database" | "table" | "view";
 export type SQLDialect = "mysql" | "postgresql";
-
-export interface ConnectionAtom {
-  parentId: InstanceId | DatabaseId | TableId | ViewId;
-  id: InstanceId | DatabaseId | TableId | ViewId;
-  key: string;
-  label: string;
-  type?: ConnectionAtomType;
-  children?: ConnectionAtom[];
-  disabled?: boolean;
-  isLeaf?: boolean;
-}
 
 export enum SortText {
   DATABASE = "0",
